Refactor Prophet Source file
